Create embeddings with the arborist package

The `embed()` function create an embedding for data set with the use of a random forest model. Two different approaches to embeddings are implemented. One is based on SHAP values while the second is a binary encoding of data flow through the forest.

embed(data, model, type = "binary")

Arguments

data

`data.frame` or `matrix` - data which will be embedded by a model

model

a tree-based model that will be used to create embeddings. If this object is an explainer, then a model will be extracted.

type

a character with the name of the embedding procedure. one out of `binary`/`shap`/`bayesian`. If `binary` then the resulting embedding will have as many dimensions as we have trees in the forest Each tree will contribute to a single dimension. The embedding procedure is described in the ,,All you need are trees'' ebook. If `shap` then Shapley values are used as embeddings. The size of the embedding corresponds to the number of variables in the 'model'.

Value

a matrix with embeddings created for observations from the 'data'

Examples

library("DALEX")
#> Warning: package ‘DALEX’ was built under R version 4.1.2
#> Welcome to DALEX (version: 2.4.2).
#> Find examples and detailed introduction at: http://ema.drwhy.ai/
model <- train(titanic_imputed, "survived", engine = "randomForest", ntree = 25)
#>  -> type set to  classification 
#>  -> leaderboard 
#>           Random Forest
#> recall    0.53125      
#> precision 0.85         
#> f1        0.6538462    
#> accuracy  0.7959184    
#> auc       0.780694     
embeddings_binary <- embed(titanic_imputed, model, type = "binary")
str(embeddings_binary)
#>  num [1:2207, 1:25] 160 176 186 156 152 ...

model_fr <- train(small_titanic, "survived", engine = "randomForest", type = "regression")
#> Error in nrow(data): object 'small_titanic' not found
embeddings_shap   <- embed(titanic_imputed, model, type = "shap")
#> Error in randomForest.unify(forest, data): Models built on data with categorical features are not supported - please encode them before training.
str(embeddings_shap)
#> Error in str(embeddings_shap): object 'embeddings_shap' not found