Extract knowledge from forester models

The `extract()` function extracts some useful statistics, like important features i.e. these used by the model, ranking of feature importances.

extract(model, ...)

Arguments

model

a model created with the `train()` function

...

other parameters

Value

set of ggplot objects that summarise the model

Examples

library(DALEX)
library(magrittr)
#> Warning: package ‘magrittr’ was built under R version 4.1.2
#> 
#> Attaching package: ‘magrittr’
#> The following object is masked from ‘package:arborist’:
#> 
#>     extract
titanic_imputed %>%
    train("survived") %>%
    forester::extract()
#> Registered S3 method overwritten by 'forester':
#>   method         from    
#>   print.forester arborist
#> [1] "This is only a placeholder - extract"