FIFA 22 preprocessed data

The fifa dataset is a preprocessed players_22.csv dataset which comes as a part of "FIFA 22 complete player dataset" at Kaggle.

data(fifa22)

Format

a data frame with 19165 rows, 62 columns and rownames

Source

The players_22.csv dataset was downloaded from the Kaggle site and went through few transformations. The complete dataset was obtained from https://www.kaggle.com/stefanoleone992/fifa-22-complete-player-dataset on August 31, 2022.

Details

It contains 19 165 players and 62 variables. Among all you will find there:

short_name (rownames)

nationality of the player (not used in modeling)

overall, potential, value_eur, wage_eur (4 potential target variables)

age, height, weight, attacking skills, defending skills, goalkeeping skills (37 variables)

It is advised to leave only one target variable for modeling.

Source: https://www.kaggle.com/datasets/stefanoleone992/fifa-22-complete-player-dataset