Train models with forester

The `train()` is the core function of this package. The only obligatory arguments are `data` and `target`. Setting and changing other arguments will affect model validation strategy, tested model families and so on.

train(
  data,
  y,
  data_test = NULL,
  y_test = NULL,
  train_test_ratio = 4,
  type = "guess",
  engine = c("randomForest", "ranger", "xgboost", "lightgbm"),
  loss = "default",
  validation = "default",
  tuning = "default",
  keep = FALSE,
  verbose = FALSE
)

Arguments

data

`data.frame` or `matrix` - data which will be used to build models. By default model will be trained on all columns in the `data`

y

target variable. It can be either (1) a vector of the same number of observations as `data` or (2) a character name of variable in the `data` that contains the target variable

data_test

data that will be used for evaluation

y_test

data that will be used for evaluation

train_test_ratio

if data_test is not provided then the orifinal data will be divided in train/test

type

a character, one of `classification`/`regression`/`guess`. sets the type of the task. If `guess` (the default option) then forester will figure out `type` based on the number of unique values in the `y` variable

engine

a vector of tree-based models that shall be testes. Possible values are: `ranger`, `xgboost`, `lightgbm`, `catboost`. All models from this vector will be trained and the best one will be returned.

loss

a character with the name of the loss function. If `default` then the default loss function for a given task will be used

validation

validation/model selection a character of data.frame. If data.frame then this data will be used for model assessment and selection. If character, then it will be a name of the hyperparameter.

tuning

a character with the name of the tuning procedure. one out of `default`/`portfolio`/`bayesian`/`random search`. If `default` then no tuning, just default hyperparameteres are used. If `portfolio` then static portfolio of hyperparametes are tried. If `bayesian` then Bayesian optimization is performed. If `random search` then random search is performed.

keep

shall all models be returned (`keep = TRUE`) or only the best one (`keep = FALSE`, default option)?

verbose

logical. If TRUE (default) then diagnostic messages will be printed

Value

an DALEX explainer (or set of explainers if keep=TRUE)

Examples

library(DALEX)
train(titanic_imputed, "survived")
#>  -> type set to  classification 
#>  -> leaderboard 
#>           Random Forest Ranger   
#> recall    0.5442177     0.4353741
#> precision 0.8333333     0.8533333
#> f1        0.6584362     0.5765766
#> accuracy  0.8117914     0.7868481
#> auc       0.7820237     0.8184668
#> Model label:  Ranger 
#> Model class:  foresterRanger,forester,ranger 
#> Data head  :
#>   gender age class    embarked  fare sibsp parch
#> 3   male  16   3rd Southampton 20.05     1     1
#> 7   male  30   2nd   Cherbourg 24.00     1     0